* Fix handling of maximise in normalise. Add tests.

+++ inst/scripts/eaf/eaf.h
@@ -174,7 +174,7 @@ static inline int percentile2level (double p, int n)
 {
     double tolerance = 1e-12;
     double x = (n * p) / 100.0;
-    double level = (x - floor(x) <= tolerance)
+    int level = (x - floor(x) <= tolerance)
         ? (int) floor(x) : (int) ceil(x);
     
     eaf_assert(level <= n);
